Modify Controller, etc..

+++ src/main/java/com/hansung/treeze/controller/MindmapController.java
@@ -5,6 +5,7 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.ModelMap;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -47,7 +48,7 @@ public class MindmapController {
 	}
 	
 @RequestMapping(value="/getMindMap/{classId}", method=RequestMethod.GET)
-	public String getMindMap(@PathVariable Long classId, ModelMap map) {
+	public String getMindMap(@PathVariable Integer classId, ModelMap map) {
 		Mindmap mindmap = mindmapService.findByclassId(classId);
 		map.put("mindmap", mindmap);
 		return "jsonView";
